pyproject.toml
[project]
name = "ask-my-biblio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "build==1.3.0",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "chromadb==1.3.2",
    "click==8.3.0",
    "coloredlogs==15.0.1",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "durationpy==0.10",
    "filelock==3.20.0",
    "flatbuffers==25.9.23",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "google-auth==2.42.1",
    "googleapis-common-protos==1.71.0",
    "greenlet==3.2.4",
    "grpcio==1.76.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.0",
    "humanfriendly==10.0",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "kubernetes==34.1.0",
    "langchain==1.0.3",
    "langchain-chroma==1.0.0",
    "langchain-classic==1.0.0",
    "langchain-community==0.4.1",
    "langchain-core==1.0.2",
    "langchain-google-genai>=3.0.1",
    "langchain-huggingface==1.0.0",
    "langchain-text-splitters==1.0.0",
    "langgraph==1.0.2",
    "langgraph-checkpoint==3.0.0",
    "langgraph-prebuilt==1.0.2",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.39",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "networkx==3.5",
    "numpy==2.3.4",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.5",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvshmem-cu12==3.3.20",
    "nvidia-nvtx-cu12==12.8.90",
    "oauthlib==3.3.1",
    "onnxruntime==1.23.2",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "orjson==3.11.4",
    "ormsgpack==1.11.0",
    "overrides==7.7.0",
    "packaging==25.0",
    "pillow==12.0.0",
    "posthog==5.4.0",
    "propcache==0.4.1",
    "protobuf==6.33.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.11.0",
    "pygments==2.19.2",
    "pypdf==6.1.3",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "referencing==0.37.0",
    "regex==2025.10.23",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rpds-py==0.28.0",
    "rsa==4.9.1",
    "safetensors==0.6.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.3",
    "sentence-transformers==5.1.2",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.44",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "threadpoolctl==3.6.0",
    "tokenizers==0.22.1",
    "torch==2.9.0",
    "tqdm==4.67.1",
    "transformers==4.57.1",
    "triton==3.5.0",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.3.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
